64-bit - not fully tested and further changes are needed

+++ include/sys/types.h
@@ -19,7 +19,7 @@
 #ifndef SYS_TYPES_H
 #define SYS_TYPES_H
 
-typedef unsigned int size_t;
-typedef signed int ssize_t;
+typedef unsigned int long long size_t;
+typedef signed int long long ssize_t;
 
 #endif
